feat: add custom hook for creating context
